+++ CHANGES.rst
@@ -5,6 +5,10 @@ Changelog
 1.8 (unreleased)
 ----------------
 
+- Allow overriding the fallback variant in img_variant_fixer.
+  Use 'medium' by default.
+  [maurits]
+
 - Let fix_html view work on the current context.  [maurits]